refactor: smtp 관련 클래스 패키징
