+++ mail-batch/src/main/java/maeilbatch/config/MailSenderConfig.java
@@ -1,5 +1,6 @@
 package maeilbatch.config;
 
+import maeilbatch.smtp.MailSenderBeanPostProcessor;
 import org.springframework.beans.factory.config.BeanPostProcessor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
